Renames package for RDA gRPC pipeline.
