+++ apps/bfd-pipeline/bfd-pipeline-rda-grpc/src/main/java/gov/cms/bfd/pipeline/rda/grpc/PreAdjudicatedClaim.java
@@ -1,4 +1,4 @@
-package gov.cms.bfd.pipeline.dc.geo;
+package gov.cms.bfd.pipeline.rda.grpc;
 
 /** Java bean representing preadjudicated claims received from RDA API. */
 public class PreAdjudicatedClaim {}
