+++ apps/bfd-pipeline/bfd-pipeline-rda-grpc/src/main/java/gov/cms/bfd/pipeline/rda/grpc/SkeletonRDASink.java
@@ -1,4 +1,4 @@
-package gov.cms.bfd.pipeline.dc.geo;
+package gov.cms.bfd.pipeline.rda.grpc;
 
 import com.codahale.metrics.Meter;
 import com.codahale.metrics.MetricRegistry;
